Create README
My Internet connection stinks, goes down all the time. I want to poll certain sites and if they are not accessible for a period of time, have Wink power off and then back on switch that the DSL modem is connected to. This will hopefully restore me connection and allow me to control cameras and thermostats from afar.